Added comment_struct for slash comment entities

+++ check_comment.h
@@ -1,2 +1,2 @@
 #include "bits/comment_struct.h" // Comment Structure: comment_s
-int check_comment(char byte, int fd, int *line_count); 
+int check_comment(char byte, int fd, long *line_count);
